+++ network/src/commonMain/kotlin/com/wire/kalium/network/BackendMetaDataUtil.kt
@@ -25,7 +25,7 @@ import com.wire.kalium.network.tools.ApiVersionDTO
 import com.wire.kalium.network.tools.ServerConfigDTO
 
 val SupportedApiVersions = setOf(0, 1, 2, 3)
-val DevelopmentApiVersions = setOf(4)
+val DevelopmentApiVersions = setOf(4, 5)
 
 interface BackendMetaDataUtil {
     fun calculateApiVersion(
