+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/base/authenticated/e2ei/E2EIApi.kt
@@ -26,7 +26,7 @@ interface E2EIApi {
     suspend fun getWireNonce(clientId: String): NetworkResponse<String>
 
     companion object {
-        fun getApiNotSupportError(apiName: String, apiVersion: String = "4") = NetworkResponse.Error(
+        fun getApiNotSupportError(apiName: String, apiVersion: String = "5") = NetworkResponse.Error(
             APINotSupported("${this::class.simpleName}: $apiName api is only available on API V$apiVersion")
         )
     }
