+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/v0/authenticated/MLSMessageApiV0.kt
@@ -26,11 +26,11 @@ import com.wire.kalium.network.utils.NetworkResponse
 internal open class MLSMessageApiV0 internal constructor() : MLSMessageApi {
 
     override suspend fun sendMessage(message: MLSMessageApi.Message): NetworkResponse<SendMLSMessageResponse> = NetworkResponse.Error(
-        APINotSupported("MLS: sendMessage api is only available on API V4")
+        APINotSupported("MLS: sendMessage api is only available on API V5")
     )
 
     override suspend fun sendCommitBundle(bundle: MLSMessageApi.CommitBundle): NetworkResponse<SendMLSMessageResponse> =
         NetworkResponse.Error(
-            APINotSupported("MLS: sendCommitBundle api is only available on API V4")
+            APINotSupported("MLS: sendCommitBundle api is only available on API V5")
         )
 }
