+++ network/src/commonTest/kotlin/com/wire/kalium/model/conversation/SubconversationDetailsResponseJson.kt
@@ -27,5 +27,5 @@ object SubconversationDetailsResponseJson {
         """.trimIndent()
     }
 
-    val v4 = ValidJsonProvider("", jsonProvider)
+    val v5 = ValidJsonProvider("", jsonProvider)
 }
